requirements.txt
streamlit
google-generativeai
requests
python-dotenv
pandas
# optional
newspaper3k
google-search-results   # for SerpAPI
